+++ src/app.ts
@@ -34,6 +34,7 @@ import {
   MAILER_CONFIG_SMTP_HOST,
   MAILER_CONFIG_USERNAME,
   PG_CONNECTION_URI,
+  PG_READ_REPLICAS_CONNECTION_URIS,
   PUBLIC_PLUGIN,
   REDIS_HOST,
   REDIS_PASSWORD,
@@ -125,8 +126,12 @@ export default async function (instance: FastifyInstance): Promise<void> {
   // load some shared schema definitions
   instance.addSchema(shared);
 
-  await instance
-    .register(fp(databasePlugin), { uri: PG_CONNECTION_URI, logs: DATABASE_LOGS })
+  instance
+    .register(fp(databasePlugin), {
+      uri: PG_CONNECTION_URI,
+      readReplicaUris: PG_READ_REPLICAS_CONNECTION_URIS,
+      logs: DATABASE_LOGS,
+    })
     .register(fp(decoratorPlugin))
     .register(fp(metaPlugin))
     .register(mailerPlugin, {