+++ src/services/itemTag/schemas/schemas.ts
@@ -136,7 +136,7 @@ const deleteOne = {
     },
   },
   response: {
-    200: { },
+    200: {},
   },
 };